+++ app/src/main/java/com/wearconnectivityexample/ui/components/AudioRecorder.kt
@@ -1,0 +1,54 @@
+package com.wearconnectivityexample.ui.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.wear.compose.material.Icon
+import androidx.wear.compose.material.Text
+import androidx.wear.compose.material.ToggleButton
+import androidx.wear.compose.material.ToggleButtonDefaults
+import com.wearconnectivityexample.R
+
+@Composable
+fun AudioRecorder(onRecordClicked: () -> Unit, checked: Boolean) {
+    val title = if (checked) "Stop Recording" else "Record Voice"
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+            Text(
+                modifier = Modifier.padding(bottom = 8.dp),
+                textAlign = TextAlign.Center,
+                color = Color.Black,
+                fontSize = 16.sp,
+                text = title,
+            )
+            ToggleButton(
+                checked = checked,
+                onCheckedChange = { onRecordClicked() },
+                enabled = true,
+            ) {
+                Icon(
+                    painter = painterResource(id = R.drawable.ic_mic),
+                    contentDescription = "airplane",
+                    modifier =
+                    Modifier
+                        .size(ToggleButtonDefaults.DefaultIconSize)
+                        .wrapContentSize(align = Alignment.Center),
+                )
+            }
+        }
+    }
+}
