+++ app/src/main/java/com/wearconnectivityexample/ui/components/ImageViewer.kt
@@ -10,7 +10,7 @@ import androidx.wear.compose.material.Button
 import androidx.wear.compose.material.Text
 import coil.compose.rememberAsyncImagePainter
 import java.io.File
-import com.wearconnectivityexample.presentation.FileState
+import com.wearconnectivityexample.data.FileState
 
 @Composable
 fun ImageViewer() {
